✨(video) add peertube pipeline for VOD
Creates a new "Peetube" pipeline for VODs. This pipeline will use the
`peertube` runners and the scaleway object storage.
Due to using scaleway object storage and not s3, no callback
will be called when the upload is finished. Therefore, the client needs
to call the 'upload-ended' API endpoint to finish the upload. Moreover,
calling this endpoint will trigger the transcoding of the VOD.
At the end of the transcoding, a callback (set in the settings) will
be called, and update the video state.
VOD uploads have their keys prefixed with `tmp`, they will be
automatically deleted by scaleway lifecycle rules.
Any "ready to use" object uploaded on the scaleway object storage
are prefixed with `scw`.
For example a video thumbnails have this format:
`scw/{video_id}/video/{stamp}/thumbnail.jpg`