♻️(back) refacto delete videos from S3
Change the way we are deleting files from S3. Before, we added a
life-cycle rule to the video directory. The problems with that is
it could result with a very long life-cycle rule list (one per video),
that we would need to update and clean up.
Now, we are moving the files to a "deleted" directory, that will have
one life-cycle rule to delete all files after a certain amount of time.
Also we are now using Celery to move files because the operation can be
long and we don't want to block the user.